+++ Components/Lists/FindFollowingList.tsx
@@ -1,0 +1,7 @@
+import SearchBar from "@/Components/Forms/SearchBar";
+
+function FindFollowingList() {
+  return <SearchBar />;
+}
+
+export default FindFollowingList;
