+++ Components/Lists/FindMemosList.tsx
@@ -1,0 +1,8 @@
+import SearchBar from "@/Components/Forms/SearchBar";
+
+
+function FindMemosList() {
+  return <SearchBar />;
+}
+
+export default FindMemosList;
