+++ Components/Lists/FindPeopleList.tsx
@@ -1,0 +1,7 @@
+import SearchBar from "@/Components/Forms/SearchBar";
+
+function FindPeopleList() {
+  return <SearchBar />;
+}
+
+export default FindPeopleList;
