+++ app/explore/page.tsx
@@ -1,9 +1,9 @@
+import ExplorerTabs from "@/Components/Tabs/ExplorerTabs";
+
 export default async function ExplorePage() {
-    return (
-        <main>
-            <h1 className='md-4 text-xl md:text-2xl'>
-                Explore
-            </h1>
-        </main>
-    )
+  return (
+    <main>
+      <ExplorerTabs />
+    </main>
+  );
 }
