+++ app/home/layout.tsx
@@ -1,5 +1,6 @@
-import SideNav from "@/app/ui/dashboard/sidenav";
 import { Metadata } from "next";
+
+import SideNav from "@/Components/Nav/SideNav";
 
 export const metadata: Metadata = {
   title: "Home",
@@ -11,7 +12,9 @@ export default function Layout({ children }: { children: React.ReactNode }) {
       <div className="w-full flex-none md:w-64">
         <SideNav />
       </div>
-      <div className="flex-grow p-6 md:overflow-y-auto md:p-12">{children}</div>
+      <div className="flex-grow p-6 pt-0 md:overflow-y-auto md:p-12 md:pt-0">
+        {children}
+      </div>
     </div>
   );
 }
