+++ app/home/page.tsx
@@ -1,13 +1,9 @@
-import Greetings from "@/app/ui/cards/greeting";
-import MemoContainer from "@/app/ui/memo/memo-container";
-import MemoList from "@/app/ui/memo-list";
+import HomeTabs from "@/Components/Tabs/HomeTabs";
 
 export default async function DashboardPage() {
-    return (
-        <main>
-            <Greetings />
-            <MemoContainer />
-            <MemoList />
-        </main>
-    )
+  return (
+    <main>
+      <HomeTabs />
+    </main>
+  );
 }
