+++ app/page.tsx
@@ -1,6 +1,6 @@
-import Link from 'next/link';
-import Image from 'next/image';
-import MemoLogo from '@/app/ui/memo-logo';
+import Link from "next/link";
+import Image from "next/image";
+import MemoLogo from "@/Components/MemoLogo";
 
 export default function LandingPage() {
   return (
@@ -12,17 +12,27 @@ export default function LandingPage() {
       </div>
       <div className="flex flex-row px-4 mt-12 justify-center items-center">
         <main>
-          <h1 className="text-black text-wrap text-[50px] font-bold p-4">Welcome to Memo</h1>
-          <p className="text-black text-wrap text-[25px] font-bold p-4">Join today.</p>
+          <h1 className="text-black text-wrap text-[50px] font-bold p-4">
+            Welcome to Memo
+          </h1>
+          <p className="text-black text-wrap text-[25px] font-bold p-4">
+            Join today.
+          </p>
           <div className="flex flex-col w-full items-center">
             <div>
-              <Link href="/signup" className="flex flex-col h-[48px] w-64 grow items-center justify-center gap-2 rounded-lg bg-blue-500 p-3 text-md text-white font-medium hover:bg-gray-600 hover:text-white outline-blue-500 md:flex-none md:justify-start md:p-2 md:px-3">
+              <Link
+                href="/signup"
+                className="flex flex-col h-[48px] w-64 grow items-center justify-center gap-2 rounded-lg bg-blue-500 p-3 text-md text-white font-medium hover:bg-gray-600 hover:text-white outline-blue-500 md:flex-none md:justify-start md:p-2 md:px-3"
+              >
                 Create account
               </Link>
             </div>
             <p className="p-6">Already have an account?</p>
             <div>
-              <Link href="/signin" className="flex flex-col h-[48px] w-64 grow items-center justify-center gap-2 rounded-lg bg-white p-3 text-md font-medium hover:bg-gray-600 hover:text-white outline-gray-600 md:flex-none md:justify-start md:p-2 md:px-3">
+              <Link
+                href="/signin"
+                className="flex flex-col h-[48px] w-64 grow items-center justify-center gap-2 rounded-lg bg-white p-3 text-md font-medium hover:bg-gray-600 hover:text-white outline-gray-600 md:flex-none md:justify-start md:p-2 md:px-3"
+              >
                 Sign in
               </Link>
             </div>
@@ -30,5 +40,5 @@ export default function LandingPage() {
         </main>
       </div>
     </div>
-  )
+  );
 }
