+++ app/profile/(overview)/page.tsx
@@ -1,15 +1,14 @@
 "use client";
 
-import ProfileCard from "@/app/ui/cards/profile-card";
+import ProfileCard from "@/Components/Cards/ProfileCard";
 
 export default function ProfilePage() {
-    
-    return (
-        <main>
-            <h1 className='md-4 text-xl md:text-2xl text-center md:text-start'>
-                Profile
-            </h1>
-            <ProfileCard />
-        </main>
-    );
+  return (
+    <main>
+      <h1 className="sticky top-0 bg-white z-10 md-4 text-xl md:text-2xl text-center md:text-start">
+        Profile
+      </h1>
+      <ProfileCard />
+    </main>
+  );
 }
